LPS-70397 Move into `auth.publicpath` package
The ideal package name would be `auth.public.path`, but as `public` is
a keyword it is an illegal package name.
