+++ modules/apps/collaboration/bookmarks/bookmarks-web/src/main/java/com/liferay/bookmarks/web/internal/auth/publicpath/AuthPublicPath.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.google.internal.web.portlet.path;
+package com.liferay.bookmarks.web.internal.auth.publicpath;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -20,7 +20,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Sergio González
  */
 @Component(
-	immediate = true, property = {"auth.public.path=/portal/google_login"},
+	immediate = true,
+	property = {
+		"auth.public.path=/bookmarks/find_entry",
+		"auth.public.path=/bookmarks/find_folder",
+		"auth.public.path=/bookmarks/open_entry"
+	},
 	service = Object.class
 )
 public class AuthPublicPath {
