+++ modules/apps/forms-and-workflow/polls/polls-web/src/main/java/com/liferay/polls/web/internal/auth/publicpath/AuthPublicPath.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.polls.web.internal.portlet.path;
+package com.liferay.polls.web.internal.auth.publicpath;
 
 import org.osgi.service.component.annotations.Component;
