+++ modules/apps/shopping/shopping-web/src/main/java/com/liferay/shopping/web/internal/auth/publicpath/AuthPublicPath.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.shopping.web.internal.portlet.path;
+package com.liferay.shopping.web.internal.auth.publicpath;
 
 import org.osgi.service.component.annotations.Component;
